+++ core_client/base/api/v3_cluster_process_get_metadata.py
@@ -1,0 +1,51 @@
+import httpx
+from ._http import execute_async, execute_sync
+from pydantic import TypeAdapter, validate_call
+
+from ...models import Client
+from ..models import Error
+from ..models.v3 import Metadata
+
+
+@validate_call()
+def _build_request(
+    client: Client,
+    id: str,
+    key: str,
+    domain: str = "",
+    retries: int = None,
+    timeout: float = None,
+):
+    if not retries:
+        retries = client.retries
+    if not timeout:
+        timeout = client.timeout
+    return {
+        "method": "get",
+        "url": f"{client.base_url}/api/v3/cluster/process/{id}/metadata/{key}?domain={domain}",
+        "headers": client.headers,
+        "timeout": timeout,
+        "data": None,
+        "json": None,
+    }, retries
+
+
+def _build_response(response: httpx.Response):
+    if response.status_code == 200:
+        response_200 = TypeAdapter(Metadata).validate_python(response.json())
+        return response_200.root
+    else:
+        response_error = TypeAdapter(Error).validate_python(response.json())
+        return response_error
+
+
+def sync(client: Client, **kwargs):
+    request, retries = _build_request(client, **kwargs)
+    response = execute_sync(client, request, retries)
+    return _build_response(response=response)
+
+
+async def asyncio(client: Client, **kwargs):
+    request, retries = _build_request(client, **kwargs)
+    response = await execute_async(client, request, retries)
+    return _build_response(response=response)