+++ core_client/base/models/v3/cluster_deployments.py
@@ -1,0 +1,13 @@
+from pydantic import BaseModel
+
+from . import ClusterDeploymentsProcess
+
+
+class ClusterDeployments(BaseModel):
+    """
+    {
+        "process": [ClusterDeploymentsProcess]
+    }
+    """
+
+    process: list[ClusterDeploymentsProcess] | None = None